created Rotating banner component parent
